+++ src/main/java/ladder/domain/Player.java
@@ -1,4 +1,4 @@
-package ladder2.domain;
+package ladder.domain;
 
 import java.util.Objects;
