+++ src/main/java/ladder/domain/PointType.java
@@ -1,4 +1,4 @@
-package ladder2.domain;
+package ladder.domain;
 
 public enum PointType {
     FIRST,
