+++ src/test/java/ladder/domain/PlayerTest.java
@@ -1,5 +1,4 @@
-package ladder2.domain;
-
+import ladder.domain.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
